Refactored SSML functions to a new file. Added a builder method for appending plain speech.
