+++ skillserver/echo.go
@@ -1,10 +1,8 @@
 package skillserver
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"time"
 )
 
@@ -80,15 +78,6 @@ func (this *EchoResponse) OutputSpeech(text string) *EchoResponse {
 	return this
 }
 
-func (this *EchoResponse) OutputSpeechSSML(text string) *EchoResponse {
-	this.Response.OutputSpeech = &EchoRespPayload{
-		Type: "SSML",
-		SSML: text,
-	}
-
-	return this
-}
-
 func (this *EchoResponse) Card(title string, content string) *EchoResponse {
 	return this.SimpleCard(title, content)
 }
@@ -133,17 +122,6 @@ func (this *EchoResponse) Reprompt(text string) *EchoResponse {
 	this.Response.Reprompt = &EchoReprompt{
 		OutputSpeech: EchoRespPayload{
 			Type: "PlainText",
-			Text: text,
-		},
-	}
-
-	return this
-}
-
-func (this *EchoResponse) RepromptSSML(text string) *EchoResponse {
-	this.Response.Reprompt = &EchoReprompt{
-		OutputSpeech: EchoRespPayload{
-			Type: "SSML",
 			Text: text,
 		},
 	}
@@ -236,79 +214,4 @@ type EchoRespPayload struct {
 	SSML    string        `json:"ssml,omitempty"`
 	Content string        `json:"content,omitempty"`
 	Image   EchoRespImage `json:"image,omitempty"`
-}
-
-// Helper Types
-
-type SSMLTextBuilder struct {
-	buffer *bytes.Buffer
-}
-
-func NewSSMLTextBuilder() *SSMLTextBuilder {
-	return &SSMLTextBuilder{bytes.NewBufferString("")}
-}
-
-func (builder *SSMLTextBuilder) AppendAmazonEffect(text, name string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<amazon:effect name=\"%s\">%s</amazon:effect>", name, text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendAudio(src string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<audio src=\"%s\"/>", src))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendBreak(text, strength, time string) *SSMLTextBuilder {
-
-	if strength == "" {
-		// The default strength is medium
-		strength = "medium"
-	}
-
-	builder.buffer.WriteString(fmt.Sprintf("<break strength=\"%s\" time=\"%s\"/>", strength, text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendEmphasis(text, level string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<emphasis level=\"%s\">%s</emphasis>", level, text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendParagraph(text string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<p>%s</p>", text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendProsody(text, rate, pitch, volume string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<prosody rate=\"%s\" pitch=\"%s\" volume=\"%s\">%s</prosody>", rate, pitch, volume, text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendSentence(text string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<s>%s</s>", text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) AppendSubstitution(text, alias string) *SSMLTextBuilder {
-
-	builder.buffer.WriteString(fmt.Sprintf("<sub alias=\"%s\">%s</sub>", alias, text))
-
-	return builder
-}
-
-func (builder *SSMLTextBuilder) Build() string {
-	return fmt.Sprintf("<speak>%s</speak>", builder.buffer.String())
 }
